Update README.md
My code runs with the correct output, but when I test it doesn't work.  Anyone notice anything wrong. 
Here is the description:

With the addition of electric cars, we have a need to create a subclass of our Car class. In this exercise, we are going to create the Electric Car subclass so that we can override the miles per gallon calculation since electric cars don’t use gallons of gas.

The Car class is complete, but you need to complete the ElectricCar class as outlined in the starter code with comments.

Once complete, use the CarTester to create both a Car and ElectricCar object and test these per the instructions in the CarTester class.